data-pjt/consumer/requirements.txt
elasticsearch==8.12.1
elastic-transport==8.17.1
psycopg2-binary==2.9.9
kafka-python==1.4.7
python-dotenv==1.0.1
httpx==0.27.0
openai==1.22.0
tiktoken==0.8.0
